Now popup shows address of current playing server.
Also a button that generates a button to copy the link to the clipoard was
added.

The gathering of the address was achieved by multiple tricks:

* Code needed to obtain the address from the server was obtained from
* Ketchup Girl here:
* http://www.ketche.net/getServerIP/

* A temporal Content Script was created as a bridge between the web page
* and the popup with an event listener that self destructs after
* execution.

* In addition to serving as bridge the Content Script also inject a
* <script> element in the page that gets the address of the current
* server, dispatchs an event and then removes itself as element.

The self destructions are to prevent memory leaks, since all elements,
events and variables are recreated everytime the user opens the popup.

The extension still needs to check the website where it was opened,
opening on anyother site will show an exception.